add helper functions to one place
