--- apps/senf-client/src/hooks/truncateString.js
@@ -1,8 +1,0 @@
-/** @format */
-
-export function truncateString(str, num) {
-  if (str.length <= num) {
-    return str;
-  }
-  return str.slice(0, num) + "...";
-}
