--- apps/senf-client/src/util/utils.js
@@ -1,11 +1,0 @@
-/**
- * Function returning the build date(as per provided epoch)
- * @param epoch Time in milliseconds
- */
-
-import moment from "moment";
- 
-export const getBuildDate = (epoch) => {
-  const buildDate = moment(epoch).format("DD-MM-YYY HH:MM");
-  return buildDate;
-};
